# pages/api-docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "4.0",
    "title": "18 Candleriggs",
    "description": "This is the Candleriggs Server based on the OpenAPI 3.0 specification.",
    "contact": {
      "name": "Superminds",
      "email": "contact@superminds.dev"
    }
  },
  "paths": {
...
    "/api/event/upload-mobile-image/{eventId}": {
      "post": {
        "tags": [
          "Events"
        ],
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Upload a event mobile image",
        "description": "This endpoint is used to upload mobile event image using eventID.\n\n__How To__\n\n1. If you dont have a eventID hit an endpoint **GET /api/event** and get a valid eventID\n2. Fill the eventID.\n3. Select image to upload and click execute\n\n**Accepted image formats**\n1. png\n\n2. jpg\n\n3. jpeg\n\n **Accepted size and dimensions (width & height)** \n\nImage size : under 200kb\n\nwidth: 280px\n\nheight: 380px",
        "operationId": "eventMobileImageUpload",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "description": "The Id of the event to upload mobile image",
            "schema": {
              "type": "string",
              "example": "63844c8bc1e969f2934f09f7"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event object",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 201
                    },
                    "msg": {
                      "type": "string",
                      "example": "Successfully uploaded"
                    },
                    "result": {
                      "allOf": [
                        {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "637487643672cb71ea1444cd"
                            }
                          }
                        },
                        {
                          "$ref": "#/components/schemas/Event"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return400"
                }
              }
            },
            "description": "Data not received from user"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return404"
                }
              }
            },
            "description": "ObjectID not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return422"
                }
              }
            },
            "description": "Validation errors"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return500"
                }
              }
            },
            "description": "Server side error."
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Delete the mobile image of an event using eventID",
        "description": "This endpoint is used to delete a mobile image of an event using eventID.\n\n__How To__\n\n1. If you dont have an eventID hit an endpoint **GET /api/event** and get a valid eventID. \n2. Fill the eventID and execute.\n\n**The below table defines the HTTP Status codes that this API may return**\n\n  | Status Code      | Description | Reason                                   |\n  | ---------------- | ------------| -----------------------------------------|\n  | 200              | OK          | Image deleted successfully.                       |\n  | 403              | FORBIDDEN | If user unauthenticated. |\n  | 404              | NOT FOUND   | If this eventId not found.              |\n  | 422              | BAD REQUEST | If the eventId is in an invalid format. |\n  | 500              | SERVER ERROR| If server error encountered.             |\n",
        "operationId": "eventMobileImageDELETE",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6399bf322f332f8babe66407"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "msg": {
                      "type": "string",
                      "example": "Deleted successfully"
                    }
                  }
                }
              }
            },
            "description": "success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return403"
                }
              }
            },
            "description": "Forbidden: If user unauthenticated."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return404"
                }
              }
            },
            "description": "eventID not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return422"
                }
              }
            },
            "description": "Invalid Id provided"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return500"
                }
              }
            },
            "description": "Server side error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Return400": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 400
          },
          "msg": {
            "type": "string",
            "example": "Bad Request: {message}"
          }
        }
      },
      "Return403": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 403
          },
          "msg": {
            "type": "string",
            "example": "Forbidden: {message}"
          }
        }
      },
      "Return404": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 404
          },
          "msg": {
            "type": "string",
            "example": "Not Found: {message}"
          }
        }
      },
...
      "Return422": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 422
          },
          "msg": {
            "type": "string",
            "example": "Unprocessable Entity"
          },
          "errors": {
            "type": "array",
            "items": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ValidationError"
                }
              ]
            }
          }
        }
      },
      "Return500": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 500
          },
          "msg": {
            "type": "string",
            "example": "Internal Server Error: {message}"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "example": "Provide Password"
          },
          "param": {
            "type": "string",
            "example": "password"
          },
          "location": {
            "type": "string",
            "example": "body"
          }
        }
      },
...
      "Event": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EventWithoutImages"
          },
          {
            "type": "object",
            "properties": {
              "desktopImage": {
                "type": "string",
                "example": "https://18-candleriggs.fra1.digitaloceanspaces.com/image1"
              },
              "mobileImage": {
                "type": "string",
                "example": "https://18-candleriggs.fra1.digitaloceanspaces.com/image2"
              }
            }
          }
        ],
        "required": [
          "name",
          "title",
          "shortDescription",
          "mainDescription",
          "link",
          "price",
          "bookingFee",
          "age",
          "eventDate",
          "eventType",
          "showStartTime",
          "doorOpeningTime",
          "displayEventListingFrom"
        ]
      },
      "EventWithoutImages": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Cabaret Show by Wild Cabaret",
            "minLength": 1,
            "maxLength": 40
          },
          "title": {
            "type": "string",
            "example": "Amazing entertainment paired with a 3-course meal",
            "minLength": 1,
            "maxLength": 80
          },
          "shortDescription": {
            "type": "string",
            "example": "Do you have a great event idea and need a space? At 18 Candleriggs we can make it happen!",
            "minLength": 1,
            "maxLength": 160
          },
          "mainDescription": {
            "type": "string",
            "example": "With a bar for adults, soft play, arts and crafts, face painting, performers and balloon artists, there is a great mix for families to let loose and have fun!",
            "minLength": 1,
            "maxLength": 500
          },
          "link": {
            "type": "string",
            "example": "https://www.fatsoma.com/e/fs8b16xn/freaky-friday",
            "minLength": 1
          },
          "price": {
            "type": "string",
            "example": 10,
            "minLength": 1
          },
          "bookingFee": {
            "type": "number",
            "example": 1,
            "minimum": 0
          },
          "age": {
            "type": "string",
            "enum": [
              "18+",
              "all ages",
              "don't show"
            ]
          },
          "eventDate": {
            "type": "string",
            "example": "2002-07-15",
            "format": "full-date",
            "externalDocs": {
              "description": "SeeMore",
              "url": "https://www.rfc-editor.org/rfc/rfc3339#section-5.6"
            }
          },
          "eventType": {
            "type": "string",
            "enum": [
              "Live",
              "Sold Out",
              "Cancelled",
              "Private Booking"
            ]
          },
          "startTime": {
            "type": "string",
            "example": "6pm",
            "minLength": 1
          },
          "doorOpeningTime": {
            "type": "string",
            "example": "8pm",
            "minLength": 1
          },
          "displayEventListingFrom": {
            "type": "string",
            "example": "2002-07-15",
            "format": "full-date",
            "externalDocs": {
              "description": "SeeMore",
              "url": "https://www.rfc-editor.org/rfc/rfc3339#section-5.6"
            }
          },
          "seoTitle": {
            "type": "string",
            "example": "Book Venue - 18 Candleriggs",
            "minLength": 1,
            "maxLength": 60
          },
          "seoDescription": {
            "type": "string",
            "example": "Are you in a mood to have fun? We are here to help. Get in touch with Us.",
            "minLength": 1,
            "maxLength": 155
          },
          "performer": {
            "type": "string",
            "example": "John doe",
            "minLength": 0
          },
          "organizer": {
            "type": "string",
            "example": "Andrew tate",
            "minLength": 0
          },
          "organizerUrl": {
            "type": "string",
            "example": "https://bestorganizer.com",
            "minLength": 0
          }
        },
        "required": [
          "name",
          "title",
          "shortDescription",
          "mainDescription",
          "link",
          "price",
          "bookingFee",
          "age",
          "eventDate",
          "eventType",
          "showStartTime",
          "doorOpeningTime",
          "displayEventListingFrom"
        ]
      },
...
    }
  }
}